pyproject.toml
[project]
name = "git-probe"
version = "0.1.0"
description = "A tool for monitoring changes in GitHub repositories"
authors = [
    {name = "Git Probe Team"}
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "datetime>=5.0",
    "python-dotenv>=1.1.0",
]
